--- tests/Browser/Modal/Components/ModalComponentDifferentEntangle.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace Tests\Browser\Modal\Components;
-
-use Livewire\Component;
-
-class ModalComponentDifferentEntangle extends Component
-{
-    public bool $test = false;
-
-    public function render(): string
-    {
-        return <<<'HTML'
-        <div>        
-            <x-modal wire="test">
-                Foo bar
-            </x-modal>
-        
-            <x-button id="open" wire:click="$toggle('test')">Open</x-button>
-        </div>
-HTML;
-    }
-}
